patch Account type used in AccountNode to_eels
