+++ cairo/tests/conftest.py
@@ -1,129 +1,8 @@
 from dataclasses import fields
 
 from dotenv import load_dotenv
-from ethereum.trace import (
-    EvmStop,
-    GasAndRefund,
-    OpEnd,
-    OpException,
-    OpStart,
-    PrecompileEnd,
-    PrecompileStart,
-    TraceEvent,
-    TransactionEnd,
-    TransactionStart,
-)
-
-from tests.utils.args_gen import Evm
 
 load_dotenv()
-
-
-def evm_trace(
-    evm: Evm,
-    event: TraceEvent,
-    trace_memory: bool = False,
-    trace_stack: bool = True,
-    trace_return_data: bool = False,
-) -> None:
-    """
-    Log the event.
-    """
-    import logging
-
-    logger = logging.getLogger("TRACE")
-
-    if isinstance(event, TransactionStart):
-        pass
-    elif isinstance(event, TransactionEnd):
-        error_name = "None" if event.error is None else event.error.__class__.__name__
-        logger.trace_eels(
-            f"TransactionEnd: gas_used: {event.gas_used}, output: {event.output}, error: {error_name}"
-        )
-    elif isinstance(event, PrecompileStart):
-        logger.trace_eels(f"PrecompileStart: {evm.message.code_address}")
-    elif isinstance(event, PrecompileEnd):
-        logger.trace_eels(f"PrecompileEnd: {evm.message.code_address}")
-    elif isinstance(event, OpStart):
-        op = event.op
-        logger.trace_eels(f"OpStart: {hex(op.value)}")
-    elif isinstance(event, OpEnd):
-        logger.trace_eels("OpEnd")
-    elif isinstance(event, OpException):
-        logger.trace_eels(f"OpException: {event.error.__class__.__name__}")
-    elif isinstance(event, EvmStop):
-        logger.trace_eels("EvmStop")
-    elif isinstance(event, GasAndRefund):
-        logger.trace_eels(f"GasAndRefund: {event.gas_cost}")
-
-
-def pytest_configure(config):
-    """
-    Global test configuration for patching core classes.
-
-    How it works:
-    1. pytest runs this hook during test collection, before any tests execute
-    2. We directly replace the class definitions in the original modules
-    3. All subsequent imports of these modules will see our patched versions
-
-    This effectively "rewrites" the module contents at the source, so whether code does:
-    from ethereum.cancun.vm import Evm
-    or:
-        import ethereum.cancun.vm
-        evm = ethereum.cancun.vm.Evm
-    They both get our mock version, because the module itself has been modified.
-    """
-    from typing import Sequence, Union
-
-    import ethereum
-    import ethereum_rlp
-    from ethereum_types.numeric import FixedUnsigned, Uint
-
-    from tests.utils.args_gen import (
-        EMPTY_ACCOUNT,
-        Account,
-        Environment,
-        Evm,
-        Message,
-        MessageCallOutput,
-        Node,
-    )
-
-    # Apply patches at module level before any tests run
-    ethereum.cancun.vm.Evm = Evm
-    ethereum.cancun.vm.Message = Message
-    ethereum.cancun.vm.Environment = Environment
-    ethereum.cancun.vm.interpreter.MessageCallOutput = MessageCallOutput
-    ethereum.cancun.fork_types.Account = Account
-    ethereum.cancun.fork_types.EMPTY_ACCOUNT = EMPTY_ACCOUNT
-
-    # TODO: Find a better way to do this?
-    # See explanation below. Lots of EELS modules import `Account` and `EMPTY_ACCOUNT` from `ethereum.cancun.fork_types`.
-    # I think these modules get loaded before this patch is applied. Thus we must replace them manually.
-    setattr(ethereum.cancun.trie, "Account", Account)
-    setattr(ethereum.cancun.state, "Account", Account)
-    setattr(ethereum.cancun.state, "EMPTY_ACCOUNT", EMPTY_ACCOUNT)
-    setattr(ethereum.cancun.fork_types, "EMPTY_ACCOUNT", EMPTY_ACCOUNT)
-    setattr(ethereum.cancun.vm.instructions.environment, "EMPTY_ACCOUNT", EMPTY_ACCOUNT)
-
-    ethereum.cancun.trie.Node = Node
-    setattr(ethereum.cancun.trie, "Node", Node)
-
-    # Mock the Extended type
-    ethereum_rlp.rlp.Extended = Union[Sequence["Extended"], bytearray, bytes, Uint, FixedUnsigned, str, bool]  # type: ignore # noqa: F821
-
-    # Patching evm_trace:
-    # - Problem: Global patches of `ethereum.trace.evm_trace` are not reflected in places where `evm_trace` is imported in EELS.
-    # - Cause: `ethereum.cancun.vm.interpreter` (and other modules) imports `evm_trace` locally (e.g., `from ethereum.trace import evm_trace`)
-    #   at module load time, caching the original `discard_evm_trace`. Patching `ethereum.trace.evm_trace` later didn’t
-    #   update this local reference due to Python’s import caching.
-    # - Solution: Explicitly patch both `ethereum.trace.evm_trace` globally and
-    #   `ethereum.cancun.vm.interpreter.evm_trace` locally (and other places where `evm_trace` is imported).
-    if config.getoption("log_cli_level") == "TRACE":
-        import ethereum.cancun.vm.interpreter
-
-        setattr(ethereum.cancun.vm.interpreter, "evm_trace", evm_trace)
-        setattr(ethereum.cancun.vm.gas, "evm_trace", evm_trace)
 
 
 def pytest_assertrepr_compare(op, left, right):
